issue-1038: merging trait Types into StructTypeUtil
