--- sdl-core/src/test/scala/io/smartdatalake/util/spark/dataset/TypesTest.scala
@@ -1,38 +1,0 @@
-/*
- * Smart Data Lake - Build your data lake the smart way.
- *
- * Copyright © 2019-2026 ELCA Informatique SA (<https://www.elca.ch>)
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-package io.smartdatalake.util.spark.dataset
-
-import org.apache.spark.sql.types._
-import org.scalatest.flatspec.AnyFlatSpec
-import org.scalatest.matchers.should.Matchers
-
-class TypesTest extends AnyFlatSpec with Matchers with Types {
-
-  "createStruct" should "created a struct" in {
-    val argument = Array[(String, DataType, Boolean)](("id", IntegerType, false),
-      ("name", StringType, false), ("birthdate", DateType, true))
-    val actual = createStruct(argument)
-    val expected = StructType(Array(StructField("id", IntegerType, nullable = false),
-      StructField("name", StringType, nullable = false),
-      StructField("birthdate", DateType, nullable = true)))
-    actual shouldEqual expected
-  }
-
-}
